# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reddit Content Platform API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Reddit Content Platform",
      "url": "https://github.com/your-username/reddit-content-platform",
      "email": "support@yourcompany.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-api-version": "1.0.0",
    "x-api-versions-supported": [
      "v1"
    ],
    "x-api-version-header": "X-API-Version",
    "x-api-media-type": "application/vnd.reddit-platform.v1+json"
  },
  "paths": {
...
    "/api/v1/blog/posts": {
      "get": {
        "tags": [
          "public-blog"
        ],
        "summary": "List Blog Posts",
        "description": "Get paginated list of published blog posts.",
        "operationId": "list_blog_posts_api_v1_blog_posts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Posts per page",
              "default": 10,
              "title": "Size"
            },
            "description": "Posts per page"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (keyword)",
              "title": "Category"
            },
            "description": "Filter by category (keyword)"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(latest|oldest|popular)$",
              "description": "Sort order",
              "default": "latest",
              "title": "Sort"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBlogPostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/blog/archive/{year}/{month}": {
      "get": {
        "tags": [
          "public-blog"
        ],
        "summary": "Get Blog Archive",
        "description": "Get blog posts for a specific month and year.",
        "operationId": "get_blog_archive_api_v1_blog_archive__year___month__get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2030,
              "minimum": 2020,
              "description": "Year",
              "title": "Year"
            },
            "description": "Year"
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "description": "Month",
              "title": "Month"
            },
            "description": "Month"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogArchiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/blog/posts/{post_id}/related": {
      "get": {
        "tags": [
          "public-blog"
        ],
        "summary": "Get Related Posts",
        "description": "Get related posts based on tags and category.",
        "operationId": "get_related_posts_api_v1_blog_posts__post_id__related_get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "description": "Number of related posts",
              "default": 5,
              "title": "Limit"
            },
            "description": "Number of related posts"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelatedPostResponse"
                  },
                  "title": "Response Get Related Posts Api V1 Blog Posts  Post Id  Related Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blog/rss": {
      "get": {
        "tags": [
          "public-blog"
        ],
        "summary": "Get Rss Feed",
        "description": "Generate RSS feed for the blog.",
        "operationId": "get_rss_feed_api_v1_blog_rss_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Number of posts in feed",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of posts in feed"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RSSFeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlogArchiveResponse": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "title": "Month"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PublicBlogPostSummary"
            },
            "type": "array",
            "title": "Posts"
          }
        },
        "type": "object",
        "required": [
          "year",
          "month",
          "count",
          "posts"
        ],
        "title": "BlogArchiveResponse",
        "description": "Schema for blog archive by date."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicBlogPostListResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PublicBlogPostSummary"
            },
            "type": "array",
            "title": "Posts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          }
        },
        "type": "object",
        "required": [
          "posts",
          "total",
          "page",
          "size",
          "pages",
          "has_next",
          "has_prev"
        ],
        "title": "PublicBlogPostListResponse",
        "description": "Schema for paginated blog post list."
      },
      "PublicBlogPostSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "meta_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "featured_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Featured Image Url"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "estimated_read_time": {
            "type": "integer",
            "title": "Estimated Read Time"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt",
            "description": "First 200 characters of content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "meta_description",
          "tags",
          "featured_image_url",
          "word_count",
          "estimated_read_time",
          "published_at",
          "excerpt"
        ],
        "title": "PublicBlogPostSummary",
        "description": "Schema for blog post summary in list views."
      },
      "RSSFeedItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "pub_date": {
            "type": "string",
            "format": "date-time",
            "title": "Pub Date"
          },
          "guid": {
            "type": "string",
            "title": "Guid"
          }
        },
        "type": "object",
        "required": [
          "title",
          "link",
          "description",
          "pub_date",
          "guid"
        ],
        "title": "RSSFeedItem",
        "description": "Schema for RSS feed item."
      },
      "RSSFeedResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "last_build_date": {
            "type": "string",
            "format": "date-time",
            "title": "Last Build Date"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RSSFeedItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "title",
          "link",
          "description",
          "language",
          "last_build_date",
          "items"
        ],
        "title": "RSSFeedResponse",
        "description": "Schema for RSS feed."
      },
...
      "RelatedPostResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "similarity_score": {
            "type": "number",
            "title": "Similarity Score",
            "description": "Similarity score based on tags and content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "excerpt",
          "published_at",
          "similarity_score"
        ],
        "title": "RelatedPostResponse",
        "description": "Schema for related posts."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://www.reddit.com/api/v1/authorize",
            "tokenUrl": "/api/v1/auth/token",
            "scopes": {
              "read": "Read Reddit posts and comments",
              "identity": "Access user identity information"
            }
          }
        },
        "description": "Reddit OAuth2 authentication"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-app.railway.app",
      "description": "Production server"
    }
  ],
...
}